% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmem-package.R
\docType{package}
\name{altmem-package}
\alias{altmem-package}
\title{altmem: Alternative Interface to cytoMEM Marker Enrichment Modeling}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
