% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmem-package.R
\docType{package}
\name{altmem-package}
\alias{altmem-package}
\title{altmem: Alternative Interface to cytoMEM Marker Enrichment Modeling}
\description{
Automatically generate and display quantitative labels for cell populations that have been identified from flow cytometry, mass cytometry, and single-cell RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igordot.github.io/altmem/}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org}

}
\keyword{internal}
